chore: rename `references upload` to `reference upload`
